mcp_client_chatbot/requirements.txt
python-dotenv>=1.0.0
requests>=2.31.0
mcp>=1.0.0
uvicorn>=0.32.1
openai>=1.68.2   